Improve performance by caching IPs in memory
Instead of reading the file on every request (very slow), the list of
IPs is loaded through various triggers and kept in memory. On each
request, it is now a much faster lookup in memory. To ensure the list
remains up to date, fsnotify is used to watch the file for changes and
reload it.